regulator: pbias: populate disable_val inorder to disable regulator
regulator_disable_regmap uses disable_val of pbias regulator descriptor
to disable the regulator. If disable_val is not populated, '0' will
be written to the register. However actual disable_val of pbias
regulator is different from '0'. Populated the disable_val in
pbias_reg_info for the various platforms and assign it to the
disable_val of pbias regulator descriptor.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>